Revert "Fixed compiler warning C4267 'conversion from 'size_t' to 'type', possible loss of data'"
Seems to have quite a few warnings in 64-bit on my machine.

BUG=angleproject:1120

This reverts commit c5cf9bc47d0ee028adbbf9e9f94ca567eec601dc.

Reviewed-on: https://chromium-review.googlesource.com/293280
Tested-by: Jamie Madill <jmadill@chromium.org>

+++ src/compiler/translator/CallDAG.cpp
@@ -71,10 +71,10 @@ class CallDAG::CallDAGCreator : public TIntermTraverser
             record.callees.reserve(data.callees.size());
             for (auto &callee : data.callees)
             {
-                record.callees.push_back(static_cast<int>(callee->index));
+                record.callees.push_back(callee->index);
             }
 
-            (*idToIndex)[data.node->getFunctionId()] = static_cast<int>(data.index);
+            (*idToIndex)[data.node->getFunctionId()] = data.index;
         }
     }
